refactor: move initialization of tree view in its own class to drastically improve the readability of the presenter
